Very first draft for Design description
This will need going over again and most likely have some edits made to
include more information about the navigation and user flow around the
website.
It's okay for now, but isn't detailed enough. I'm going to do a bit of
work on the prototype and then come back to it to make the necessary
revisions.