Submitted
There is an option to resubmit, but I think this will be it. I'm not
really happy with it, in honesty. I feel I could have done much better,
should I have put the time into it. I say that, but I was also working
hard on the other tasked assignments and this one became a lower
priority to those rewarding 100% module marks to the module.

It's okay. It's not amazing. It's okay.